+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Changelog
 
 ## Unreleased
+- Reduce token account cache size
+- Do not throw exception for setting channel on Kin 4 payments
+- Check for duplicate signers for Stellar transactions
 
 ## [0.4.3](https://github.com/kinecosystem/kin-python/releases/tag/0.4.3)
 - Fix uploaded wheel to remove `agora/model/keys.py`